+++ src/CcgSalesApi/CcgSalesApi.php
@@ -2,27 +2,37 @@
 
 namespace Nexusvc\CcgSalesApi;
 
+use Nexusvc\CcgSalesApi\Traits\Configable;
 use Nexusvc\CcgSalesApi\Traits\Jsonable;
 
 class CcgSalesApi {
 
     use Jsonable;
+    use Configable;
 
-    protected $config;
+    protected $env;
 
-    protected $crypt = Crypt\Crypt::class;
-
-    protected $env = 'development';
+    protected $crypt    = Crypt\Crypt::class;
 
     public $applicant   = Applicant\Applicant::class;
+
     public $auth        = Auth\Authentication::class;
+
     public $client      = Client\Client::class;
+
     public $order       = Order\Order::class;
+
     public $payable     = Payable\Payable::class;
+
     public $quote       = Quote\Quote::class;
 
     public function __construct() {
-        static::boot();
+                
+        $this->loadConfigs();
+        $this->setEnvironment();
+        
+        $this->auth();
+        
         $this->order = new Order\Order;
         $this->crypt = new Crypt\Crypt;
     }
@@ -35,27 +45,19 @@ class CcgSalesApi {
         return static::${$method};
     }
 
-    public function boot() {
-        $this->loadConfigs();
-        $this->auth();
-    }
-
     public function auth() {
-
         return $this->auth = ( $this->auth instanceof Auth\Authentication ) ? 
             $this->auth : 
             new Auth\Authentication;
     }
 
     public function crypt() {
-
         return $this->crypt = ( $this->crypt instanceof Crypt ) ? 
             $this->crypt : 
             new Crypt;
     }
 
     public function client() {
-
         return $this->client = ( $this->client instanceof Client\Client ) ? 
             $this->client : 
             new Client\Client;
@@ -70,7 +72,6 @@ class CcgSalesApi {
     }
 
     public function quote(array $params = []) {
-
         return $this->quote = ( $this->quote instanceof Quote\Quote ) ? 
             new Quote\Quote($this->auth(), $params ) : 
             new Quote\Quote($this->auth(), $params );
@@ -78,61 +79,6 @@ class CcgSalesApi {
 
     public static function getInstance() {
         return new self();
-    }
-
-    public static function url($endpoint = null) {
-        $ccg = static::getInstance();
-        $endpoints = array_dot($ccg->config['endpoints']);
-        $suffix = array_has($endpoints, $endpoint) ? $endpoints[$endpoint] : $endpoint;
-
-        return $endpoints['base.'.$ccg->env] . $suffix;
-    }
-
-    protected function loadConfigs() {
-        $this->config = collect([]);
-        $dir = new \DirectoryIterator(dirname(__FILE__).'/Config');
-        foreach ($dir as $fileinfo) {
-            if (!$fileinfo->isDot()) {
-                $this->mergeConfigFrom(
-                    __DIR__.'/Config/'.$fileinfo->getFilename(), 
-                    str_replace('.php','',$fileinfo->getFilename())
-                );
-            }
-        }
-
-        return $this;
-    }
-
-    public static function config($key = null, $dot = true) {
-        $ccg = static::getInstance();
-        
-        $dot_config = array_dot($ccg->config);
-
-        if(!$dot) {
-            return (is_null($key) ? 
-                $ccg->config : 
-                ( array_has($ccg->config, $key) ? 
-                    $ccg->config[$key] : 
-                    null
-                )
-            );
-        }
-
-        return (is_null($key) ? 
-            $dot_config : 
-            ( array_has($dot_config, $key) ? 
-                $dot_config[$key] : 
-                ( array_has($ccg->config, $key) ? 
-                    $ccg->config[$key] : 
-                    null 
-                ) 
-            ) 
-        );
-    }
-
-    protected function mergeConfigFrom($path, $key) {
-        $config = $this->config->get($key, []);
-        $this->config->put($key, array_merge(require $path, $config));
     }
 
 }
