+++ src/CcgSalesApi/Crypt/Crypt.php
@@ -3,8 +3,6 @@
 namespace Nexusvc\CcgSalesApi\Crypt;
 
 use Illuminate\Encryption\Encrypter;
-use Tightenco\Collect\Support\Collection;
-use Tightenco\Collect\Support\Arr;
 
 class Crypt extends Encrypter {
 
@@ -19,20 +17,23 @@ class Crypt extends Encrypter {
      */
     public function __construct($key = null, $cipher = 'AES-256-CBC')
     {
-        
         $this->loadConfigs();
-
         if(is_null($key)) $key = $this->key;
-
         if(strpos($key, ':') !== false && explode(':', $key)[0] === 'base64') {
             $key = base64_decode(explode(':', $key)[1]);
         }
-
         parent::__construct($key, $cipher);
     }
 
+    public static function config($key = null, $dot = true) {
+        $ccg = static::getInstance();
+        $dot_config = array_dot($ccg->config);
+        if(!$dot) return (is_null($key) ? $ccg->config : (Arr::has($ccg->config, $key) ? $ccg->config[$key] : null));
+        return (is_null($key) ? $dot_config : (Arr::has($dot_config, $key) ? $dot_config[$key] : (Arr::has($ccg->config, $key) ? $ccg->config[$key] : null ) ) );
+    }
+
     protected function loadConfigs() {
-        $this->config = new Collection;
+        $this->config = collect([]);
         $dir = new \DirectoryIterator(dirname(__FILE__).'/../Config');
         foreach ($dir as $fileinfo) {
             if (!$fileinfo->isDot()) {
@@ -43,8 +44,6 @@ class Crypt extends Encrypter {
             }
         }
 
-        // Filter Config to Encrpytion
-        // $cryptConfigs = collect([]);
         try {
             $this->key = $this->config['env']['encryption']['key'];
             $this->cipher = $this->config['env']['encryption']['cipher'];
@@ -53,16 +52,6 @@ class Crypt extends Encrypter {
         }
         
         return $this;
-    }
-
-    public static function config($key = null, $dot = true) {
-        $ccg = static::getInstance();
-        
-        $dot_config = Arr::dot($ccg->config);
-
-        if(!$dot) return (is_null($key) ? $ccg->config : (Arr::has($ccg->config, $key) ? $ccg->config[$key] : null));
-
-        return (is_null($key) ? $dot_config : (Arr::has($dot_config, $key) ? $dot_config[$key] : (Arr::has($ccg->config, $key) ? $ccg->config[$key] : null ) ) );
     }
 
     protected function mergeConfigFrom($path, $key)
