+++ src/CcgSalesApi/Schema/Schema.php
@@ -9,6 +9,7 @@ class Schema {
     use Jsonable;
 
     protected $instance;
+    
     protected $payload;
 
     protected $formatted = [];
